+++ app.js
@@ -16,8 +16,6 @@ app.use(logger(formatsLogger));
 app.use(cors());
 app.use(express.json());
 
-app.use(express.json());
-
 app.use("/avatars", avatarRouter);
 
 app.use("/api/users", usersRouter);
