+++ controllers/usersController.js
@@ -5,7 +5,7 @@ const {
   login,
   logout,
   changeSubscription,
-} = require("../models/users");
+} = require("../models/users/index");
 
 const signupController = async (req, res, next) => {
   const { email, password } = req.body;
