+++ middlewares/addValidation.js
@@ -1,5 +1,4 @@
 const Joi = require("joi");
-const { ValidationError } = "../helpers/errors.js";
 
 const addValidation = (req, res, next) => {
   const schema = Joi.object({
@@ -21,19 +20,4 @@ const addValidation = (req, res, next) => {
   next();
 };
 
-const patchValidation = (req, res, next) => {
-  const schema = Joi.object({
-    favorite: Joi.boolean().required(),
-  });
-
-  const validationResult = schema.validate(req.body);
-  if (validationResult.error) {
-    next(new ValidationError(JSON.stringify(validationResult.error.details)));
-  }
-  next();
-};
-
-module.exports = {
-  addValidation,
-  patchValidation,
-};
+module.exports = addValidation;
