+++ middlewares/authMiddleware.js
@@ -24,6 +24,4 @@ const authMiddleware = async (req, res, next) => {
   }
 };
 
-module.exports = {
-  authMiddleware,
-};
+module.exports = authMiddleware;
