+++ middlewares/index.js
@@ -1,0 +1,13 @@
+const authMiddleware = require("./authMiddleware");
+const loginValidation = require("./loginValidation");
+const validateSubscriptionUpdate = require("./validateSubscriptionUpdate");
+const addValidation = require("./addValidation");
+const patchValidation = require("./patchValidation");
+
+module.exports = {
+  authMiddleware,
+  loginValidation,
+  validateSubscriptionUpdate,
+  addValidation,
+  patchValidation,
+};
