+++ models/contacts/patchContact.js
@@ -1,0 +1,9 @@
+const { Contact } = require("../../db/contactModel");
+
+const patchContact = async (id, { favorite }) => {
+  await Contact.findByIdAndUpdate(id, {
+    $set: { favorite },
+  });
+};
+
+module.exports = patchContact;
