+++ models/contacts/removeContact.js
@@ -1,0 +1,7 @@
+const { Contact } = require("../../db/contactModel");
+
+const removeContact = async (id) => {
+  await Contact.findByIdAndRemove(id);
+};
+
+module.exports = removeContact;
