+++ models/users/index.js
@@ -1,0 +1,11 @@
+const signup = require("./signup");
+const login = require("./login");
+const logout = require("./logout");
+const changeSubscription = require("./changeSubscription");
+
+module.exports = {
+  signup,
+  login,
+  logout,
+  changeSubscription,
+};
